CP-27111: implement QMP device_add command according to specification
The original device_add code was closely attached to adding USB devices,
and wouldn't allow the addition of other types of devices.

https://github.com/qemu/qemu/blob/master/qapi-schema.json#L1478 defines
the QMP command device_add. This command accepts an arbitrary number of
key-value pairs for an unlimited number of different types of devices,
and only requires the 'driver' key to be present.

This patch detaches the device_add handling from the USB device implementation,
making it possible to add different types of devices in the future. There is no
semantic change for the existing USB device implementation.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>