CP-29936: add query-migratable QMP command
This will be useful to check that a VM is migratable before reaching the
point of no return during migration.
E.g. VMs that have NVME devices are not migratable, but they become
migratable when the PV drivers detach them.

It is safer to ask QEMU whether it is migratable than to rely on the
presence of PV drivers to determine this (could be that you have PV
        drivers but the device failed to unplug for some reason).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>