CP-30136: fix PCI device add
There is no dev and fn in the JSON protocol, only addr

Now that https://github.com/xapi-project/xenopsd/pull/620 is close to
working I was able to test this and Qemu rejected the QMP command.

The proper way to specify the device:function for the guest is to use
`addr` as seen here:
https://github.com/mirage/xen/blob/7c2c176/tools/libxl/libxl_qmp.c#L866-L869

With the updated QMP command Qemu no longer rejects the command.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>